Size of types

+++ 0x00-hello_world/6-size.c
@@ -12,7 +12,7 @@ int main(void)
 
 	long c;
 
-	long int d;
+	long long int d;
 
 	float e;
